refactor: isolate prompt cache snapshots
