+++ scripts/rust-source-size-baseline.txt
@@ -7,7 +7,7 @@ crates/agent-engine/src/config.rs 2556
 crates/agent-engine/src/policy.rs 1177
 crates/agent-engine/src/rollout.rs 2266
 crates/agent-engine/src/runtime/engine.rs 2816
-crates/agent-engine/src/runtime/prompt_cache.rs 2233
+crates/agent-engine/src/runtime/prompt_cache.rs 1617
 crates/agent-engine/src/runtime/submission_handlers.rs 2981
 crates/agent-engine/src/runtime/tool_orchestrator.rs 3291
 crates/agent-engine/src/runtime/tool_policy.rs 1107
